requirementss.txt
asgiref==3.2.10
azt==0.0.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudinary==1.22.0
Cython==0.29.21
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.3.0
django-jet==1.0.8
django-rest-multiple-models==2.1.3
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gunicorn==20.0.4
idna==2.10
Pillow==7.2.0
psycopg2==2.8.5
pyfiglet==0.8.post1
PyJWT==1.7.1
pytz==2020.1
pyzt==1.0.1
requests==2.24.0
Serializer==0.2.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
urllib3==1.25.10
whitenoise==5.2.0
